refactor: change the structure of files and folders
