+++ src/components/FolderInfo/FolderInfo.css
