+++ src/components/Footer/SnsLogo.jsx
