+++ src/components/index.js
@@ -1,0 +1,11 @@
+export { default as App } from "./App";
+export { default as CardSection } from "./Card/CardSection";
+export { default as Card } from "./Card/Card";
+export { default as FolderInfo } from "./FolderInfo/FolderInfo";
+export { default as Footer } from "./Footer/Footer";
+export { default as SnsCollection } from "./Footer/SnsCollection";
+export { default as SnsLogo } from "./Footer/SnsLogo";
+export { default as LoginButton } from "./Navigator/LoginButton";
+export { default as Logo } from "./Navigator/Logo";
+export { default as Navigator } from "./Navigator/NavigationBar";
+export { default as SearchBar } from "./SearchBar/SearchBar";
