+++ src/index.jsx
@@ -1,6 +1,6 @@
 import ReactDOM from "react-dom/client";
-import App from "./components/App";
-import "./style/index.css";
+import Main from "Main";
+import "style/index.css";
 
 const root = ReactDOM.createRoot(document.getElementById("root"));
-root.render(<App />);
+root.render(<Main />);
