+++ src/pages/SharedPage.css
@@ -1,0 +1,31 @@
+/* PC: 1200px 이상
+Tablet: 768px 이상 ~ 1199px 이하
+Mobile: 375px 이상 ~ 767px 이하 */
+
+header {
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  gap: 2rem;
+  padding: 3rem 0 6rem;
+  background-color: #edf7ff;
+}
+
+article {
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  gap: 4rem;
+  padding-top: 4rem;
+  padding-bottom: 5rem;
+
+  @media (max-width: 1199px) {
+    padding: 4rem 3rem 4rem;
+  }
+
+  @media (max-width: 767px) {
+    padding: 3rem 3rem 4rem;
+  }
+}
